New build

+++ aubit4glsrc/trunk/tools/no_yacc/cygwin/compilers/4glc/mk_states_1.h
@@ -1,4 +1,4 @@
-const short *token_groups[]={
+const int *token_groups[]={
     token_group_0,
     token_group_1,
     token_group_2,
